@@ -15,7 +15,7 @@
 	},
 	"require": {
 		"psr/cache": "^1.0",
-		"php": "^7.0"
+		"php": "^7.4 || ^8.0"
 	},
 	"require-dev": {
 		"predis/predis": "^1.1",
